wasm poc: Add cmake support for wasm32-wasi targets
Modify CMakeLists to build for wasm32-wasi. This requires
external dependencies wasi-sdk, openssl-wasm, and zstd-wasm.
Additionally, disable all tests as we do not have a wasm build
for googletest or benchmark.

The current SIMD instructions for wasm are stubbed out and need
to be completed by including the wasm_simd128.h file defined
in the upstream LLVM repository.

After installing with make install, the libmdoc_static.a library
is an archive of wasm bytecode objects and can be linked with other
wasm sources.