wasm poc: Lay groundwork for supporting emscripten target
Update the link directory at the beginning for the target
ABI (WASI or emscripten). This requires modifications to
the parent library, so this will not be merged into the
poc branch until that is ready.

Additionally, I still need to figure out how to make use
of the emscripten toolchain file.